Make more space in the status table in README.md.
Move Octavo to a footnote until more benchmarks are implemented. Some
other libraries have had fewer benchmarks submitted, but there is
active progress in changing that for many of them. Plus, this
conveniently puts the table in an order that's more alphabetical and
that isn't biased towards *ring*, which wasn't my intent.